.history/requirements_20250303140206.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
gunicorn==21.2.0
python-logging-loki==0.3.1